+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/main/java/org/hypertrace/traceenricher/enrichment/enrichers/backend/provider/KafkaBackendProvider.java
@@ -1,4 +1,4 @@
-package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
+package org.hypertrace.traceenricher.enrichment.enrichers.backend.provider;
 
 import java.util.Collections;
 import java.util.Map;
@@ -8,24 +8,24 @@ import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
 import org.hypertrace.semantic.convention.utils.messaging.MessagingSemanticConventionUtils;
 import org.hypertrace.traceenricher.enrichment.enrichers.BackendType;
 
-public class SqsBackendResolver extends AbstractBackendResolver {
-  public SqsBackendResolver(FqnResolver fqnResolver) {
-    super(fqnResolver);
-  }
+public class KafkaBackendProvider implements BackendProvider {
+
+  @Override
+  public void init(Event event) {}
 
   @Override
   public boolean isValidBackend(Event event) {
-    return MessagingSemanticConventionUtils.isSqsBackend(event);
+    return MessagingSemanticConventionUtils.isKafkaBackend(event);
   }
 
   @Override
   public BackendType getBackendType(Event event) {
-    return BackendType.SQS;
+    return BackendType.KAFKA;
   }
 
   @Override
   public Optional<String> getBackendUri(Event event, StructuredTraceGraph structuredTraceGraph) {
-    return MessagingSemanticConventionUtils.getSqsBackendURI(event);
+    return MessagingSemanticConventionUtils.getKafkaBackendURI(event);
   }
 
   @Override
@@ -41,6 +41,6 @@ public class SqsBackendResolver extends AbstractBackendResolver {
 
   @Override
   public Optional<String> getBackendDestination(Event event) {
-    return MessagingSemanticConventionUtils.getMessagingDestination(event);
+    return MessagingSemanticConventionUtils.getMessagingDestinationForKafka(event);
   }
 }
