+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/main/java/org/hypertrace/traceenricher/enrichment/enrichers/backend/provider/SqsBackendProvider.java
@@ -1,31 +1,30 @@
-package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
+package org.hypertrace.traceenricher.enrichment.enrichers.backend.provider;
 
 import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
 import org.hypertrace.core.datamodel.Event;
 import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
-import org.hypertrace.semantic.convention.utils.db.DbSemanticConventionUtils;
+import org.hypertrace.semantic.convention.utils.messaging.MessagingSemanticConventionUtils;
 import org.hypertrace.traceenricher.enrichment.enrichers.BackendType;
 
-public class CassandraBackendResolver extends AbstractBackendResolver {
-  public CassandraBackendResolver(FqnResolver fqnResolver) {
-    super(fqnResolver);
-  }
+public class SqsBackendProvider implements BackendProvider {
+  @Override
+  public void init(Event event) {}
 
   @Override
   public boolean isValidBackend(Event event) {
-    return DbSemanticConventionUtils.isCassandraBackend(event);
+    return MessagingSemanticConventionUtils.isSqsBackend(event);
   }
 
   @Override
   public BackendType getBackendType(Event event) {
-    return BackendType.CASSANDRA;
+    return BackendType.SQS;
   }
 
   @Override
   public Optional<String> getBackendUri(Event event, StructuredTraceGraph structuredTraceGraph) {
-    return DbSemanticConventionUtils.getCassandraURI(event);
+    return MessagingSemanticConventionUtils.getSqsBackendURI(event);
   }
 
   @Override
@@ -36,11 +35,11 @@ public class CassandraBackendResolver extends AbstractBackendResolver {
 
   @Override
   public Optional<String> getBackendOperation(Event event) {
-    return DbSemanticConventionUtils.getDbOperation(event);
+    return MessagingSemanticConventionUtils.getMessagingOperation(event);
   }
 
   @Override
   public Optional<String> getBackendDestination(Event event) {
-    return DbSemanticConventionUtils.getDestinationForCassandra(event);
+    return MessagingSemanticConventionUtils.getMessagingDestination(event);
   }
 }
