+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/test/java/org/hypertrace/traceenricher/enrichment/enrichers/backend/provider/CassandraBackendProviderTest.java
@@ -1,10 +1,13 @@
-package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
+package org.hypertrace.traceenricher.enrichment.enrichers.backend.provider;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 
+import com.typesafe.config.Config;
+import com.typesafe.config.ConfigFactory;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
+import java.util.List;
 import java.util.Map;
 import org.hypertrace.core.datamodel.AttributeValue;
 import org.hypertrace.core.datamodel.Attributes;
@@ -16,73 +19,71 @@ import org.hypertrace.core.datamodel.Metrics;
 import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
 import org.hypertrace.core.datamodel.shared.trace.AttributeValueCreator;
 import org.hypertrace.entity.data.service.v1.Entity;
+import org.hypertrace.traceenricher.enrichment.clients.ClientRegistry;
+import org.hypertrace.traceenricher.enrichment.enrichers.backend.AbstractBackendEntityEnricher;
+import org.hypertrace.traceenricher.enrichment.enrichers.backend.FqnResolver;
+import org.hypertrace.traceenricher.enrichment.enrichers.backend.HypertraceFqnResolver;
+import org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend.BackendInfo;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public class ElasticsearchBackendResolverTest {
-  private FqnResolver fqnResolver;
-  private ElasticSearchBackendResolver elasticsSarchBackendResolver;
+public class CassandraBackendProviderTest {
+
+  private AbstractBackendEntityEnricher backendEntityEnricher;
   private StructuredTraceGraph structuredTraceGraph;
 
   @BeforeEach
   public void setup() {
-    fqnResolver = new HypertraceFqnResolver();
-    elasticsSarchBackendResolver = new ElasticSearchBackendResolver(fqnResolver);
+    backendEntityEnricher = new MockBackendEntityEnricher();
+    backendEntityEnricher.init(ConfigFactory.empty(), mock(ClientRegistry.class));
+
     structuredTraceGraph = mock(StructuredTraceGraph.class);
   }
 
   @Test
   public void testBackendResolution() {
     BackendInfo backendInfo =
-        elasticsSarchBackendResolver.resolve(getElasticsearchEvent(), structuredTraceGraph).get();
+        backendEntityEnricher.resolve(getCassandraEvent(), structuredTraceGraph).get();
     Entity entity = backendInfo.getEntity();
-    Assertions.assertEquals("test-index", entity.getEntityName());
+    Assertions.assertEquals("localhost:9000", entity.getEntityName());
     Map<String, AttributeValue> attributes = backendInfo.getAttributes();
     assertEquals(
         Map.of(
             "BACKEND_OPERATION",
-            AttributeValueCreator.create("GET"),
+            AttributeValueCreator.create("select"),
             "BACKEND_DESTINATION",
-            AttributeValueCreator.create("test-index")),
+            AttributeValueCreator.create("customer.orders")),
         attributes);
   }
 
   @Test
   public void testBackendResolutionForOTEvent() {
     BackendInfo backendInfo =
-        elasticsSarchBackendResolver.resolve(getElasticsearchOTEvent(), structuredTraceGraph).get();
+        backendEntityEnricher.resolve(getCassandraOTEvent(), structuredTraceGraph).get();
     Entity entity = backendInfo.getEntity();
-    Assertions.assertEquals("test", entity.getEntityName());
+    Assertions.assertEquals("localhost:9000", entity.getEntityName());
     Map<String, AttributeValue> attributes = backendInfo.getAttributes();
-    assertEquals(
-        Map.of(
-            "BACKEND_OPERATION",
-            AttributeValueCreator.create("GET"),
-            "BACKEND_DESTINATION",
-            AttributeValueCreator.create("test-index")),
-        attributes);
+    assertEquals(Map.of("BACKEND_OPERATION", AttributeValueCreator.create("select")), attributes);
   }
 
   @Test
-  public void testBackendResolutionForOTelEvent() {
+  public void testBackendResolutionWithoutConnectionString() {
     BackendInfo backendInfo =
-        elasticsSarchBackendResolver
-            .resolve(getElasticsearchOTelEvent(), structuredTraceGraph)
-            .get();
+        backendEntityEnricher.resolve(getCassandraOTelEvent(), structuredTraceGraph).get();
     Entity entity = backendInfo.getEntity();
-    Assertions.assertEquals("test:2000", entity.getEntityName());
+    Assertions.assertEquals("test:9000", entity.getEntityName());
     Map<String, AttributeValue> attributes = backendInfo.getAttributes();
     assertEquals(
         Map.of(
             "BACKEND_OPERATION",
-            AttributeValueCreator.create("GetAction"),
+            AttributeValueCreator.create("select"),
             "BACKEND_DESTINATION",
-            AttributeValueCreator.create("test-index")),
+            AttributeValueCreator.create("customer.orders")),
         attributes);
   }
 
-  private Event getElasticsearchEvent() {
+  private Event getCassandraEvent() {
     Event event =
         Event.newBuilder()
             .setCustomerId("__default")
@@ -97,20 +98,24 @@ public class ElasticsearchBackendResolverTest {
                 Attributes.newBuilder()
                     .setAttributeMap(
                         Map.of(
-                            "elasticsearch.url",
-                            AttributeValue.newBuilder().setValue("test-index").build(),
+                            "db.connection_string",
+                            AttributeValue.newBuilder()
+                                .setValue("jdbc:cassandra://localhost:9000")
+                                .build(),
                             "span.kind",
                             AttributeValue.newBuilder().setValue("client").build(),
                             "db.system",
-                            AttributeValue.newBuilder().setValue("elasticsearch").build(),
+                            AttributeValue.newBuilder().setValue("cassandra").build(),
+                            "db.statement",
+                            AttributeValue.newBuilder()
+                                .setValue("select * from audit_message")
+                                .build(),
                             "db.operation",
-                            AttributeValue.newBuilder().setValue("GET").build(),
-                            "elasticsearch.method",
-                            AttributeValue.newBuilder().setValue("PUT").build(),
-                            "elasticsearch.action",
-                            AttributeValue.newBuilder().setValue("GetAction").build(),
-                            "elasticsearch.request.indices",
-                            AttributeValue.newBuilder().setValue("test-index").build(),
+                            AttributeValue.newBuilder().setValue("select").build(),
+                            "db.cassandra.table",
+                            AttributeValue.newBuilder().setValue("orders").build(),
+                            "db.cassandra.keyspace",
+                            AttributeValue.newBuilder().setValue("customer").build(),
                             "k8s.pod_id",
                             AttributeValue.newBuilder()
                                 .setValue("55636196-c840-11e9-a417-42010a8a0064")
@@ -123,7 +128,7 @@ public class ElasticsearchBackendResolverTest {
                             "FLAGS",
                             AttributeValue.newBuilder().setValue("0").build()))
                     .build())
-            .setEventName("elasticsearch.connection")
+            .setEventName("cassandra.connection")
             .setStartTimeMillis(1566869077746L)
             .setEndTimeMillis(1566869077750L)
             .setMetrics(
@@ -142,7 +147,7 @@ public class ElasticsearchBackendResolverTest {
     return event;
   }
 
-  private Event getElasticsearchOTEvent() {
+  private Event getCassandraOTEvent() {
     Event event =
         Event.newBuilder()
             .setCustomerId("__default")
@@ -157,20 +162,18 @@ public class ElasticsearchBackendResolverTest {
                 Attributes.newBuilder()
                     .setAttributeMap(
                         Map.of(
-                            "peer.service",
-                            AttributeValue.newBuilder().setValue("elasticsearch").build(),
                             "peer.hostname",
-                            AttributeValue.newBuilder().setValue("test").build(),
+                            AttributeValue.newBuilder().setValue("localhost").build(),
+                            "peer.port",
+                            AttributeValue.newBuilder().setValue("9000").build(),
                             "span.kind",
                             AttributeValue.newBuilder().setValue("client").build(),
                             "db.system",
-                            AttributeValue.newBuilder().setValue("elasticsearch").build(),
-                            "db.operation",
-                            AttributeValue.newBuilder().setValue("GET").build(),
-                            "elasticsearch.method",
-                            AttributeValue.newBuilder().setValue("PUT").build(),
-                            "elasticsearch.request.indices",
-                            AttributeValue.newBuilder().setValue("test-index").build(),
+                            AttributeValue.newBuilder().setValue("cassandra").build(),
+                            "db.statement",
+                            AttributeValue.newBuilder()
+                                .setValue("select * from audit_message")
+                                .build(),
                             "k8s.pod_id",
                             AttributeValue.newBuilder()
                                 .setValue("55636196-c840-11e9-a417-42010a8a0064")
@@ -183,7 +186,7 @@ public class ElasticsearchBackendResolverTest {
                             "FLAGS",
                             AttributeValue.newBuilder().setValue("0").build()))
                     .build())
-            .setEventName("elasticsearch.connection")
+            .setEventName("cassandra.connection")
             .setStartTimeMillis(1566869077746L)
             .setEndTimeMillis(1566869077750L)
             .setMetrics(
@@ -202,7 +205,7 @@ public class ElasticsearchBackendResolverTest {
     return event;
   }
 
-  private Event getElasticsearchOTelEvent() {
+  private Event getCassandraOTelEvent() {
     Event event =
         Event.newBuilder()
             .setCustomerId("__default")
@@ -217,18 +220,22 @@ public class ElasticsearchBackendResolverTest {
                 Attributes.newBuilder()
                     .setAttributeMap(
                         Map.of(
-                            "net.peer.port",
-                            AttributeValue.newBuilder().setValue("2000").build(),
                             "net.peer.name",
                             AttributeValue.newBuilder().setValue("test").build(),
+                            "net.peer.port",
+                            AttributeValue.newBuilder().setValue("9000").build(),
                             "span.kind",
                             AttributeValue.newBuilder().setValue("client").build(),
                             "db.system",
-                            AttributeValue.newBuilder().setValue("elasticsearch").build(),
-                            "elasticsearch.action",
-                            AttributeValue.newBuilder().setValue("GetAction").build(),
-                            "elasticsearch.request.indices",
-                            AttributeValue.newBuilder().setValue("test-index").build(),
+                            AttributeValue.newBuilder().setValue("cassandra").build(),
+                            "db.statement",
+                            AttributeValue.newBuilder()
+                                .setValue("select * from audit_message")
+                                .build(),
+                            "db.cassandra.table",
+                            AttributeValue.newBuilder().setValue("orders").build(),
+                            "db.cassandra.keyspace",
+                            AttributeValue.newBuilder().setValue("customer").build(),
                             "k8s.pod_id",
                             AttributeValue.newBuilder()
                                 .setValue("55636196-c840-11e9-a417-42010a8a0064")
@@ -241,7 +248,7 @@ public class ElasticsearchBackendResolverTest {
                             "FLAGS",
                             AttributeValue.newBuilder().setValue("0").build()))
                     .build())
-            .setEventName("elasticsearch.connection")
+            .setEventName("cassandra.connection")
             .setStartTimeMillis(1566869077746L)
             .setEndTimeMillis(1566869077750L)
             .setMetrics(
@@ -258,5 +265,21 @@ public class ElasticsearchBackendResolverTest {
                         .build()))
             .build();
     return event;
+  }
+
+  static class MockBackendEntityEnricher extends AbstractBackendEntityEnricher {
+
+    @Override
+    public void setup(Config enricherConfig, ClientRegistry clientRegistry) {}
+
+    @Override
+    public List<BackendProvider> getBackendProviders() {
+      return List.of(new CassandraBackendProvider());
+    }
+
+    @Override
+    public FqnResolver getFqnResolver() {
+      return new HypertraceFqnResolver();
+    }
   }
 }
