+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/test/java/org/hypertrace/traceenricher/enrichment/enrichers/backend/provider/HttpBackendProviderTest.java
@@ -1,10 +1,13 @@
-package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
+package org.hypertrace.traceenricher.enrichment.enrichers.backend.provider;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 
+import com.typesafe.config.Config;
+import com.typesafe.config.ConfigFactory;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
+import java.util.List;
 import java.util.Map;
 import org.hypertrace.core.datamodel.AttributeValue;
 import org.hypertrace.core.datamodel.Attributes;
@@ -20,20 +23,25 @@ import org.hypertrace.core.span.constants.v1.Http;
 import org.hypertrace.entity.constants.v1.BackendAttribute;
 import org.hypertrace.entity.data.service.v1.Entity;
 import org.hypertrace.traceenricher.enrichedspan.constants.v1.Backend;
+import org.hypertrace.traceenricher.enrichment.clients.ClientRegistry;
+import org.hypertrace.traceenricher.enrichment.enrichers.backend.AbstractBackendEntityEnricher;
 import org.hypertrace.traceenricher.enrichment.enrichers.BackendType;
+import org.hypertrace.traceenricher.enrichment.enrichers.backend.FqnResolver;
+import org.hypertrace.traceenricher.enrichment.enrichers.backend.HypertraceFqnResolver;
+import org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend.BackendInfo;
 import org.hypertrace.traceenricher.util.Constants;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public class HttpBackendResolverTest {
-  private FqnResolver fqnResolver;
-  private HttpBackendResolver backendResolver;
+public class HttpBackendProviderTest {
+  private AbstractBackendEntityEnricher backendEntityEnricher;
   private StructuredTraceGraph structuredTraceGraph;
 
   @BeforeEach
   public void setup() {
-    this.fqnResolver = new HypertraceFqnResolver();
-    backendResolver = new HttpBackendResolver(this.fqnResolver);
+    backendEntityEnricher = new MockBackendEntityEnricher();
+    backendEntityEnricher.init(ConfigFactory.empty(), mock(ClientRegistry.class));
+
     structuredTraceGraph = mock(StructuredTraceGraph.class);
   }
 
@@ -103,7 +111,7 @@ public class HttpBackendResolverTest {
                     .build())
             .build();
 
-    final BackendInfo backendInfo = backendResolver.resolve(e, structuredTraceGraph).get();
+    final BackendInfo backendInfo = backendEntityEnricher.resolve(e, structuredTraceGraph).get();
     final Entity backendEntity = backendInfo.getEntity();
     assertEquals(backendEntity.getEntityName(), "dataservice:9394");
     assertEquals(3, backendEntity.getIdentifyingAttributesCount());
@@ -224,7 +232,8 @@ public class HttpBackendResolverTest {
                     .build())
             .build();
 
-    final Entity backendEntity = backendResolver.resolve(e, structuredTraceGraph).get().getEntity();
+    final Entity backendEntity =
+        backendEntityEnricher.resolve(e, structuredTraceGraph).get().getEntity();
     assertEquals("dataservice:9394", backendEntity.getEntityName());
     assertEquals(3, backendEntity.getIdentifyingAttributesCount());
     assertEquals(
@@ -329,7 +338,8 @@ public class HttpBackendResolverTest {
                     .build())
             .build();
 
-    final Entity backendEntity = backendResolver.resolve(e, structuredTraceGraph).get().getEntity();
+    final Entity backendEntity =
+        backendEntityEnricher.resolve(e, structuredTraceGraph).get().getEntity();
     assertEquals("dataservice:9394", backendEntity.getEntityName());
     assertEquals(3, backendEntity.getIdentifyingAttributesCount());
     assertEquals(
@@ -435,7 +445,8 @@ public class HttpBackendResolverTest {
                     .build())
             .build();
 
-    final Entity backendEntity = backendResolver.resolve(e, structuredTraceGraph).get().getEntity();
+    final Entity backendEntity =
+        backendEntityEnricher.resolve(e, structuredTraceGraph).get().getEntity();
     assertEquals(backendEntity.getEntityName(), "dataservice:9394");
     assertEquals(3, backendEntity.getIdentifyingAttributesCount());
     assertEquals(
@@ -550,7 +561,7 @@ public class HttpBackendResolverTest {
                     .build())
             .build();
 
-    Entity backendEntity = backendResolver.resolve(e, structuredTraceGraph).get().getEntity();
+    Entity backendEntity = backendEntityEnricher.resolve(e, structuredTraceGraph).get().getEntity();
     assertEquals("dataservice:9394", backendEntity.getEntityName());
     assertEquals(3, backendEntity.getIdentifyingAttributesCount());
     assertEquals(
@@ -668,7 +679,8 @@ public class HttpBackendResolverTest {
                     .build())
             .build();
 
-    final Entity backendEntity = backendResolver.resolve(e, structuredTraceGraph).get().getEntity();
+    final Entity backendEntity =
+        backendEntityEnricher.resolve(e, structuredTraceGraph).get().getEntity();
     assertEquals(backendEntity.getEntityName(), "dataservice:9394");
     assertEquals(3, backendEntity.getIdentifyingAttributesCount());
     assertEquals(
@@ -713,5 +725,21 @@ public class HttpBackendResolverTest {
             .getValue()
             .getString(),
         "GET");
+  }
+
+  static class MockBackendEntityEnricher extends AbstractBackendEntityEnricher {
+
+    @Override
+    public void setup(Config enricherConfig, ClientRegistry clientRegistry) {}
+
+    @Override
+    public List<BackendProvider> getBackendProviders() {
+      return List.of(new HttpBackendProvider());
+    }
+
+    @Override
+    public FqnResolver getFqnResolver() {
+      return new HypertraceFqnResolver();
+    }
   }
 }
