+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/test/java/org/hypertrace/traceenricher/enrichment/enrichers/backend/provider/RabbitMqBackendProviderTest.java
@@ -1,10 +1,13 @@
-package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
+package org.hypertrace.traceenricher.enrichment.enrichers.backend.provider;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 
+import com.typesafe.config.Config;
+import com.typesafe.config.ConfigFactory;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
+import java.util.List;
 import java.util.Map;
 import org.hypertrace.core.datamodel.AttributeValue;
 import org.hypertrace.core.datamodel.Attributes;
@@ -16,70 +19,83 @@ import org.hypertrace.core.datamodel.Metrics;
 import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
 import org.hypertrace.core.datamodel.shared.trace.AttributeValueCreator;
 import org.hypertrace.entity.data.service.v1.Entity;
+import org.hypertrace.traceenricher.enrichment.clients.ClientRegistry;
+import org.hypertrace.traceenricher.enrichment.enrichers.backend.AbstractBackendEntityEnricher;
+import org.hypertrace.traceenricher.enrichment.enrichers.backend.FqnResolver;
+import org.hypertrace.traceenricher.enrichment.enrichers.backend.HypertraceFqnResolver;
+import org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend.BackendInfo;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public class CassandraBackendResolverTest {
-
-  private FqnResolver fqnResolver;
-  private CassandraBackendResolver cassandraBackendResolver;
+/** Unit test for {@link RabbitMqBackendProvider} */
+public class RabbitMqBackendProviderTest {
+  private AbstractBackendEntityEnricher backendEntityEnricher;
   private StructuredTraceGraph structuredTraceGraph;
 
   @BeforeEach
   public void setup() {
-    fqnResolver = new HypertraceFqnResolver();
-    cassandraBackendResolver = new CassandraBackendResolver(fqnResolver);
+    backendEntityEnricher = new MockBackendEntityEnricher();
+    backendEntityEnricher.init(ConfigFactory.empty(), mock(ClientRegistry.class));
+
     structuredTraceGraph = mock(StructuredTraceGraph.class);
   }
 
   @Test
-  public void testBackendResolution() {
+  public void testEventResolution() {
+    String routingKey = "routingkey";
     BackendInfo backendInfo =
-        cassandraBackendResolver.resolve(getCassandraEvent(), structuredTraceGraph).get();
+        backendEntityEnricher.resolve(getRabbitMqEvent(routingKey), structuredTraceGraph).get();
     Entity entity = backendInfo.getEntity();
-    Assertions.assertEquals("localhost:9000", entity.getEntityName());
+    Assertions.assertEquals(routingKey, entity.getEntityName());
     Map<String, AttributeValue> attributes = backendInfo.getAttributes();
     assertEquals(
         Map.of(
             "BACKEND_OPERATION",
-            AttributeValueCreator.create("select"),
+            AttributeValueCreator.create("receive"),
             "BACKEND_DESTINATION",
-            AttributeValueCreator.create("customer.orders")),
+            AttributeValueCreator.create("routingkey.QueueName")),
         attributes);
   }
 
   @Test
-  public void testBackendResolutionForOTEvent() {
+  public void testBackendOperationAndDestinationResolution() {
+    String routingKey = "routingkey";
     BackendInfo backendInfo =
-        cassandraBackendResolver.resolve(getCassandraOTEvent(), structuredTraceGraph).get();
+        backendEntityEnricher
+            .resolve(getRabbitMqEventMissingOperation(routingKey), structuredTraceGraph)
+            .get();
     Entity entity = backendInfo.getEntity();
-    Assertions.assertEquals("localhost:9000", entity.getEntityName());
-    Map<String, AttributeValue> attributes = backendInfo.getAttributes();
-    assertEquals(Map.of("BACKEND_OPERATION", AttributeValueCreator.create("select")), attributes);
-  }
-
-  @Test
-  public void testBackendResolutionWithoutConnectionString() {
-    BackendInfo backendInfo =
-        cassandraBackendResolver.resolve(getCassandraOTelEvent(), structuredTraceGraph).get();
-    Entity entity = backendInfo.getEntity();
-    Assertions.assertEquals("test:9000", entity.getEntityName());
+    Assertions.assertEquals(routingKey, entity.getEntityName());
     Map<String, AttributeValue> attributes = backendInfo.getAttributes();
     assertEquals(
         Map.of(
             "BACKEND_OPERATION",
-            AttributeValueCreator.create("select"),
+            AttributeValueCreator.create("basic.publish"),
             "BACKEND_DESTINATION",
-            AttributeValueCreator.create("customer.orders")),
+            AttributeValueCreator.create("routingkey.QueueName")),
         attributes);
   }
 
-  private Event getCassandraEvent() {
+  @Test
+  public void testBackendDestinationResolutionWithRoutingKey() {
+    String routingKey = "routingkey";
+    BackendInfo backendInfo =
+        backendEntityEnricher
+            .resolve(getRabbitMqDestinationWithRoutingKey(routingKey), structuredTraceGraph)
+            .get();
+    Entity entity = backendInfo.getEntity();
+    Assertions.assertEquals(routingKey, entity.getEntityName());
+    Map<String, AttributeValue> attributes = backendInfo.getAttributes();
+    assertEquals(
+        Map.of("BACKEND_DESTINATION", AttributeValueCreator.create("routingkey")), attributes);
+  }
+
+  private Event getRabbitMqEvent(String routingKey) {
     Event event =
         Event.newBuilder()
-            .setCustomerId("__default")
-            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f8".getBytes()))
+            .setCustomerId("customer1")
+            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f9".getBytes()))
             .setEntityIdList(Arrays.asList("4bfca8f7-4974-36a4-9385-dd76bf5c8824"))
             .setEnrichedAttributes(
                 Attributes.newBuilder()
@@ -90,24 +106,18 @@ public class CassandraBackendResolverTest {
                 Attributes.newBuilder()
                     .setAttributeMap(
                         Map.of(
-                            "db.connection_string",
+                            "rabbitmq.routing_key",
+                            AttributeValue.newBuilder().setValue(routingKey).build(),
+                            "rabbitmq.message",
                             AttributeValue.newBuilder()
-                                .setValue("jdbc:cassandra://localhost:9000")
+                                .setValue("updating user's last session")
                                 .build(),
                             "span.kind",
                             AttributeValue.newBuilder().setValue("client").build(),
-                            "db.system",
-                            AttributeValue.newBuilder().setValue("cassandra").build(),
-                            "db.statement",
-                            AttributeValue.newBuilder()
-                                .setValue("select * from audit_message")
-                                .build(),
-                            "db.operation",
-                            AttributeValue.newBuilder().setValue("select").build(),
-                            "db.cassandra.table",
-                            AttributeValue.newBuilder().setValue("orders").build(),
-                            "db.cassandra.keyspace",
-                            AttributeValue.newBuilder().setValue("customer").build(),
+                            "messaging.operation",
+                            AttributeValue.newBuilder().setValue("receive").build(),
+                            "messaging.destination",
+                            AttributeValue.newBuilder().setValue("QueueName").build(),
                             "k8s.pod_id",
                             AttributeValue.newBuilder()
                                 .setValue("55636196-c840-11e9-a417-42010a8a0064")
@@ -120,7 +130,7 @@ public class CassandraBackendResolverTest {
                             "FLAGS",
                             AttributeValue.newBuilder().setValue("0").build()))
                     .build())
-            .setEventName("cassandra.connection")
+            .setEventName("rabbitmq.connection")
             .setStartTimeMillis(1566869077746L)
             .setEndTimeMillis(1566869077750L)
             .setMetrics(
@@ -139,11 +149,11 @@ public class CassandraBackendResolverTest {
     return event;
   }
 
-  private Event getCassandraOTEvent() {
+  private Event getRabbitMqEventMissingOperation(String routingKey) {
     Event event =
         Event.newBuilder()
-            .setCustomerId("__default")
-            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f8".getBytes()))
+            .setCustomerId("customer1")
+            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f9".getBytes()))
             .setEntityIdList(Arrays.asList("4bfca8f7-4974-36a4-9385-dd76bf5c8824"))
             .setEnrichedAttributes(
                 Attributes.newBuilder()
@@ -154,18 +164,18 @@ public class CassandraBackendResolverTest {
                 Attributes.newBuilder()
                     .setAttributeMap(
                         Map.of(
-                            "peer.hostname",
-                            AttributeValue.newBuilder().setValue("localhost").build(),
-                            "peer.port",
-                            AttributeValue.newBuilder().setValue("9000").build(),
+                            "rabbitmq.routing_key",
+                            AttributeValue.newBuilder().setValue(routingKey).build(),
+                            "rabbitmq.message",
+                            AttributeValue.newBuilder()
+                                .setValue("updating user's last session")
+                                .build(),
                             "span.kind",
                             AttributeValue.newBuilder().setValue("client").build(),
-                            "db.system",
-                            AttributeValue.newBuilder().setValue("cassandra").build(),
-                            "db.statement",
-                            AttributeValue.newBuilder()
-                                .setValue("select * from audit_message")
-                                .build(),
+                            "rabbitmq.command",
+                            AttributeValue.newBuilder().setValue("basic.publish").build(),
+                            "messaging.destination",
+                            AttributeValue.newBuilder().setValue("QueueName").build(),
                             "k8s.pod_id",
                             AttributeValue.newBuilder()
                                 .setValue("55636196-c840-11e9-a417-42010a8a0064")
@@ -178,7 +188,7 @@ public class CassandraBackendResolverTest {
                             "FLAGS",
                             AttributeValue.newBuilder().setValue("0").build()))
                     .build())
-            .setEventName("cassandra.connection")
+            .setEventName("rabbitmq.connection")
             .setStartTimeMillis(1566869077746L)
             .setEndTimeMillis(1566869077750L)
             .setMetrics(
@@ -197,11 +207,11 @@ public class CassandraBackendResolverTest {
     return event;
   }
 
-  private Event getCassandraOTelEvent() {
+  private Event getRabbitMqDestinationWithRoutingKey(String routingKey) {
     Event event =
         Event.newBuilder()
-            .setCustomerId("__default")
-            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f8".getBytes()))
+            .setCustomerId("customer1")
+            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f9".getBytes()))
             .setEntityIdList(Arrays.asList("4bfca8f7-4974-36a4-9385-dd76bf5c8824"))
             .setEnrichedAttributes(
                 Attributes.newBuilder()
@@ -212,35 +222,18 @@ public class CassandraBackendResolverTest {
                 Attributes.newBuilder()
                     .setAttributeMap(
                         Map.of(
-                            "net.peer.name",
-                            AttributeValue.newBuilder().setValue("test").build(),
-                            "net.peer.port",
-                            AttributeValue.newBuilder().setValue("9000").build(),
+                            "rabbitmq.routing_key",
+                            AttributeValue.newBuilder().setValue(routingKey).build(),
+                            "rabbitmq.message",
+                            AttributeValue.newBuilder()
+                                .setValue("updating user's last session")
+                                .build(),
                             "span.kind",
                             AttributeValue.newBuilder().setValue("client").build(),
-                            "db.system",
-                            AttributeValue.newBuilder().setValue("cassandra").build(),
-                            "db.statement",
-                            AttributeValue.newBuilder()
-                                .setValue("select * from audit_message")
-                                .build(),
-                            "db.cassandra.table",
-                            AttributeValue.newBuilder().setValue("orders").build(),
-                            "db.cassandra.keyspace",
-                            AttributeValue.newBuilder().setValue("customer").build(),
-                            "k8s.pod_id",
-                            AttributeValue.newBuilder()
-                                .setValue("55636196-c840-11e9-a417-42010a8a0064")
-                                .build(),
-                            "docker.container_id",
-                            AttributeValue.newBuilder()
-                                .setValue(
-                                    "ee85cf2cfc3b24613a3da411fdbd2f3eabbe729a5c86c5262971c8d8c29dad0f")
-                                .build(),
                             "FLAGS",
                             AttributeValue.newBuilder().setValue("0").build()))
                     .build())
-            .setEventName("cassandra.connection")
+            .setEventName("rabbitmq.connection")
             .setStartTimeMillis(1566869077746L)
             .setEndTimeMillis(1566869077750L)
             .setMetrics(
@@ -257,5 +250,21 @@ public class CassandraBackendResolverTest {
                         .build()))
             .build();
     return event;
+  }
+
+  static class MockBackendEntityEnricher extends AbstractBackendEntityEnricher {
+
+    @Override
+    public void setup(Config enricherConfig, ClientRegistry clientRegistry) {}
+
+    @Override
+    public List<BackendProvider> getBackendProviders() {
+      return List.of(new RabbitMqBackendProvider());
+    }
+
+    @Override
+    public FqnResolver getFqnResolver() {
+      return new HypertraceFqnResolver();
+    }
   }
 }
