--- hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/test/java/org/hypertrace/traceenricher/enrichment/enrichers/resolver/backend/KafkaBackendResolverTest.java
@@ -1,209 +1,0 @@
-package org.hypertrace.traceenricher.enrichment.enrichers.resolver.backend;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.mock;
-
-import java.nio.ByteBuffer;
-import java.util.Arrays;
-import java.util.Map;
-import org.hypertrace.core.datamodel.AttributeValue;
-import org.hypertrace.core.datamodel.Attributes;
-import org.hypertrace.core.datamodel.Event;
-import org.hypertrace.core.datamodel.EventRef;
-import org.hypertrace.core.datamodel.EventRefType;
-import org.hypertrace.core.datamodel.MetricValue;
-import org.hypertrace.core.datamodel.Metrics;
-import org.hypertrace.core.datamodel.shared.StructuredTraceGraph;
-import org.hypertrace.core.datamodel.shared.trace.AttributeValueCreator;
-import org.hypertrace.entity.data.service.v1.Entity;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-/** Unit Test for {@link KafkaBackendResolver} */
-public class KafkaBackendResolverTest {
-
-  private FqnResolver fqnResolver;
-  private KafkaBackendResolver kafkaBackendResolver;
-  private StructuredTraceGraph structuredTraceGraph;
-
-  @BeforeEach
-  public void setup() {
-    fqnResolver = new HypertraceFqnResolver();
-    kafkaBackendResolver = new KafkaBackendResolver(fqnResolver);
-    structuredTraceGraph = mock(StructuredTraceGraph.class);
-  }
-
-  @Test
-  public void TestOtelBackendEventResolution() {
-    String broker = "kafka-test.hypertrace.com:9092";
-    BackendInfo backendInfo =
-        kafkaBackendResolver.resolve(getOtelKafkaBackendEvent(broker), structuredTraceGraph).get();
-    Entity entity = backendInfo.getEntity();
-    Assertions.assertEquals(broker, entity.getEntityName());
-    Map<String, AttributeValue> attributes = backendInfo.getAttributes();
-    assertEquals(Map.of("BACKEND_OPERATION", AttributeValueCreator.create("receive")), attributes);
-  }
-
-  @Test
-  public void TestOTBackendEventResolution() {
-    String brokerHost = "kafka-test.hypertrace.com";
-    String brokerPort = "9092";
-    Entity entity =
-        kafkaBackendResolver
-            .resolve(getOTKafkaBackendEvent(brokerHost, brokerPort), structuredTraceGraph)
-            .get()
-            .getEntity();
-    Assertions.assertEquals(String.format("%s:%s", brokerHost, brokerPort), entity.getEntityName());
-  }
-
-  @Test
-  public void TestOtelBackendDestinationResolution() {
-    String broker = "kafka-test.hypertrace.com:9092";
-    BackendInfo backendInfo =
-        kafkaBackendResolver
-            .resolve(getOtelKafkaBackendEventForDestination(broker), structuredTraceGraph)
-            .get();
-    Entity entity = backendInfo.getEntity();
-    Assertions.assertEquals(broker, entity.getEntityName());
-    Map<String, AttributeValue> attributes = backendInfo.getAttributes();
-    assertEquals(
-        Map.of("BACKEND_DESTINATION", AttributeValueCreator.create("myGroup.QueueName")),
-        attributes);
-  }
-
-  private Event getOtelKafkaBackendEvent(String broker) {
-    Event event =
-        Event.newBuilder()
-            .setCustomerId("customer1")
-            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f9".getBytes()))
-            .setEntityIdList(Arrays.asList("4bfca8f7-4974-36a4-9385-dd76bf5c8824"))
-            .setEnrichedAttributes(
-                Attributes.newBuilder()
-                    .setAttributeMap(
-                        Map.of("SPAN_TYPE", AttributeValue.newBuilder().setValue("EXIT").build()))
-                    .build())
-            .setAttributes(
-                Attributes.newBuilder()
-                    .setAttributeMap(
-                        Map.of(
-                            "messaging.system",
-                            AttributeValue.newBuilder().setValue("kafka").build(),
-                            "messaging.url",
-                            AttributeValue.newBuilder().setValue(broker).build(),
-                            "span.kind",
-                            AttributeValue.newBuilder().setValue("client").build(),
-                            "messaging.operation",
-                            AttributeValue.newBuilder().setValue("receive").build(),
-                            "FLAGS",
-                            AttributeValue.newBuilder().setValue("0").build()))
-                    .build())
-            .setEventName("kafka.connection")
-            .setStartTimeMillis(1566869077746L)
-            .setEndTimeMillis(1566869077750L)
-            .setMetrics(
-                Metrics.newBuilder()
-                    .setMetricMap(
-                        Map.of("Duration", MetricValue.newBuilder().setValue(4.0).build()))
-                    .build())
-            .setEventRefList(
-                Arrays.asList(
-                    EventRef.newBuilder()
-                        .setTraceId(ByteBuffer.wrap("random_trace_id".getBytes()))
-                        .setEventId(ByteBuffer.wrap("random_event_id".getBytes()))
-                        .setRefType(EventRefType.CHILD_OF)
-                        .build()))
-            .build();
-
-    return event;
-  }
-
-  private Event getOTKafkaBackendEvent(String host, String port) {
-    Event event =
-        Event.newBuilder()
-            .setCustomerId("customer1")
-            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f9".getBytes()))
-            .setEntityIdList(Arrays.asList("4bfca8f7-4974-36a4-9385-dd76bf5c8824"))
-            .setEnrichedAttributes(
-                Attributes.newBuilder()
-                    .setAttributeMap(
-                        Map.of("SPAN_TYPE", AttributeValue.newBuilder().setValue("EXIT").build()))
-                    .build())
-            .setAttributes(
-                Attributes.newBuilder()
-                    .setAttributeMap(
-                        Map.of(
-                            "peer.service", AttributeValue.newBuilder().setValue("kafka").build(),
-                            "peer.hostname", AttributeValue.newBuilder().setValue(host).build(),
-                            "peer.port", AttributeValue.newBuilder().setValue(port).build(),
-                            "span.kind", AttributeValue.newBuilder().setValue("client").build(),
-                            "FLAGS", AttributeValue.newBuilder().setValue("0").build()))
-                    .build())
-            .setEventName("kafka.connection")
-            .setStartTimeMillis(1566869077746L)
-            .setEndTimeMillis(1566869077750L)
-            .setMetrics(
-                Metrics.newBuilder()
-                    .setMetricMap(
-                        Map.of("Duration", MetricValue.newBuilder().setValue(4.0).build()))
-                    .build())
-            .setEventRefList(
-                Arrays.asList(
-                    EventRef.newBuilder()
-                        .setTraceId(ByteBuffer.wrap("random_trace_id".getBytes()))
-                        .setEventId(ByteBuffer.wrap("random_event_id".getBytes()))
-                        .setRefType(EventRefType.CHILD_OF)
-                        .build()))
-            .build();
-
-    return event;
-  }
-
-  private Event getOtelKafkaBackendEventForDestination(String broker) {
-    Event event =
-        Event.newBuilder()
-            .setCustomerId("customer1")
-            .setEventId(ByteBuffer.wrap("bdf03dfabf5c70f9".getBytes()))
-            .setEntityIdList(Arrays.asList("4bfca8f7-4974-36a4-9385-dd76bf5c8824"))
-            .setEnrichedAttributes(
-                Attributes.newBuilder()
-                    .setAttributeMap(
-                        Map.of("SPAN_TYPE", AttributeValue.newBuilder().setValue("EXIT").build()))
-                    .build())
-            .setAttributes(
-                Attributes.newBuilder()
-                    .setAttributeMap(
-                        Map.of(
-                            "messaging.system",
-                            AttributeValue.newBuilder().setValue("kafka").build(),
-                            "messaging.url",
-                            AttributeValue.newBuilder().setValue(broker).build(),
-                            "span.kind",
-                            AttributeValue.newBuilder().setValue("client").build(),
-                            "messaging.destination",
-                            AttributeValue.newBuilder().setValue("QueueName").build(),
-                            "messaging.kafka.consumer_group",
-                            AttributeValue.newBuilder().setValue("myGroup").build(),
-                            "FLAGS",
-                            AttributeValue.newBuilder().setValue("0").build()))
-                    .build())
-            .setEventName("kafka.connection")
-            .setStartTimeMillis(1566869077746L)
-            .setEndTimeMillis(1566869077750L)
-            .setMetrics(
-                Metrics.newBuilder()
-                    .setMetricMap(
-                        Map.of("Duration", MetricValue.newBuilder().setValue(4.0).build()))
-                    .build())
-            .setEventRefList(
-                Arrays.asList(
-                    EventRef.newBuilder()
-                        .setTraceId(ByteBuffer.wrap("random_trace_id".getBytes()))
-                        .setEventId(ByteBuffer.wrap("random_event_id".getBytes()))
-                        .setRefType(EventRefType.CHILD_OF)
-                        .build()))
-            .build();
-
-    return event;
-  }
-}
