+++ hypertrace-trace-enricher/hypertrace-trace-enricher-impl/src/test/resources/enricher.conf
@@ -32,7 +32,7 @@ enricher {
   }
 
   BackendEntityEnricher {
-    class = "org.hypertrace.traceenricher.enrichment.enrichers.BackendEntityEnricher"
+    class = "org.hypertrace.traceenricher.enrichment.enrichers.backend.BackendEntityEnricher"
     entity.service.config = {
       host = localhost
       port = 50061
