+++ hypertrace-trace-enricher/hypertrace-trace-enricher/src/main/resources/configs/common/application.conf
@@ -75,7 +75,7 @@ enricher {
   }
 
   BackendEntityEnricher {
-    class = "org.hypertrace.traceenricher.enrichment.enrichers.BackendEntityEnricher"
+    class = "org.hypertrace.traceenricher.enrichment.enrichers.backend.BackendEntityEnricher"
     dependencies = ["DefaultServiceEntityEnricher", "EndpointEnricher"]
   }
