+++ hypertrace-trace-enricher/hypertrace-trace-enricher/src/test/resources/configs/hypertrace-trace-enricher/application.conf
@@ -64,7 +64,7 @@ enricher {
   }
 
   BackendEntityEnricher {
-    class = "org.hypertrace.traceenricher.enrichment.enrichers.BackendEntityEnricher"
+    class = "org.hypertrace.traceenricher.enrichment.enrichers.backend.BackendEntityEnricher"
     dependencies = ["DefaultServiceEntityEnricher"]
   }
